bot/requirements.txt
git+https://github.com/Gorialis/jishaku
discord.py
asyncpg
asyncpg-stubs
python-dotenv